# openapi.yaml
openapi: 3.0.0
info:
  title: Zombie DnD API
  version: 1.0.0
  description: API persistente per un'avventura D&D ambientata nel mondo di The Walking Dead

servers:
  - url: https://zombie-dnd-api.onrender.com

paths:
  /personaggi/:
    post:
      summary: Crea un nuovo personaggio
      operationId: creaPersonaggio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Personaggio'
      responses:
        '200':
          description: Personaggio creato
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                  id:
                    type: string
    get:
      summary: Elenco personaggi
      operationId: listaPersonaggi
      responses:
        '200':
          description: Lista personaggi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Personaggio'

  /personaggi/{id}:
    get:
      summary: Recupera un personaggio
      operationId: getPersonaggio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dettaglio personaggio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Personaggio'

  /relazione/:
    post:
      summary: Crea o aggiorna una relazione
      operationId: aggiornaRelazione
      parameters:
        - in: query
          name: id_da
          required: true
          schema:
            type: string
        - in: query
          name: id_a
          required: true
          schema:
            type: string
        - in: query
          name: tipo
          required: true
          schema:
            type: string
        - in: query
          name: fiducia
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Risposta
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string

  /relazione/fiducia/:
    put:
      summary: Modifica la fiducia di una relazione
      operationId: modificaFiducia
      parameters:
        - in: query
          name: id_da
          required: true
          schema:
            type: string
        - in: query
          name: id_a
          required: true
          schema:
            type: string
        - in: query
          name: variazione
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Risposta
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string

  /dado/:
    get:
      summary: Lancia un dado
      operationId: tiraDado
      parameters:
        - in: query
          name: lati
          required: false
          schema:
            type: integer
        - in: query
          name: modificatore
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Risultato
          content:
            application/json:
              schema:
                type: object
                properties:
                  tiro:
                    type: integer
                  modificatore:
                    type: integer
                  totale:
                    type: integer
                  esito:
                    type: string

  /duello/:
    get:
      summary: Tiro contrapposto
      operationId: duello
      parameters:
        - in: query
          name: mod_utente
          schema:
            type: integer
        - in: query
          name: mod_nemico
          schema:
            type: integer
      responses:
        '200':
          description: Risultato
          content:
            application/json:
              schema:
                type: object
                properties:
                  utente:
                    type: integer
                  nemico:
                    type: integer
                  esito:
                    type: string

  /azione/:
    post:
      summary: Azione narrativa
      operationId: azioneNarrativa
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: tipo
          required: true
          schema:
            type: string
        - in: query
          name: abilita
          required: true
          schema:
            type: string
        - in: query
          name: difficolta
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Risultato azione
          content:
            application/json:
              schema:
                type: object
                properties:
                  tiro:
                    type: integer
                  totale:
                    type: integer
                  esito:
                    type: string
                  critico:
                    type: string
                  effetti:
                    type: array
                    items:
                      type: string

  /evento-casuale/:
    get:
      summary: Genera evento casuale
      operationId: eventoCasuale
      responses:
        '200':
          description: Evento
          content:
            application/json:
              schema:
                type: object
                properties:
                  tipo:
                    type: string
                  evento:
                    type: string

  /luoghi/:
    get:
      summary: Elenco luoghi
      operationId: elencoLuoghi
      responses:
        '200':
          description: Mappa
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string

  /spostamento/:
    post:
      summary: Sposta personaggio
      operationId: spostaPersonaggio
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: destinazione
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Risultato spostamento
          content:
            application/json:
              schema:
                type: object
                properties:
                  personaggio:
                    type: string
                  da:
                    type: string
                  a:
                    type: string
                  evento:
                    type: string
                  salute_attuale:
                    type: integer

components:
  schemas:
    Personaggio:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        eta:
          type: integer
        ruolo:
          type: string
        lingua:
          type: string
        stato:
          type: string
        salute:
          type: integer
        sanita_mentale:
          type: integer
        inventario:
          type: object
          additionalProperties:
            type: integer
        relazioni:
          type: object
          additionalProperties:
            type: object
            properties:
              tipo:
                type: string
              fiducia:
                type: integer
        competenze:
          type: array
          items:
            type: string
        ultimo_luogo:
          type: string